requirements.txt
kafka-python==2.0.2
python-dotenv==1.0.0
faker==20.1.0
psycopg2-binary==2.9.9
redis==5.0.1
